refactor: move categoryDTOs to new package
