+++ rentplace/src/main/java/kattsyn/dev/rentplace/controllers/CategoryController.java
@@ -9,9 +9,9 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
-import kattsyn.dev.rentplace.dtos.CategoryCreateEditDTO;
-import kattsyn.dev.rentplace.dtos.CategoryDTO;
-import kattsyn.dev.rentplace.dtos.ImageDTO;
+import kattsyn.dev.rentplace.dtos.categories.CategoryCreateEditDTO;
+import kattsyn.dev.rentplace.dtos.categories.CategoryDTO;
+import kattsyn.dev.rentplace.dtos.images.ImageDTO;
 import kattsyn.dev.rentplace.entities.Image;
 import kattsyn.dev.rentplace.services.CategoryService;
 import lombok.RequiredArgsConstructor;
