+++ rentplace/src/main/java/kattsyn/dev/rentplace/dtos/categories/CategoryCreateEditDTO.java
@@ -1,4 +1,4 @@
-package kattsyn.dev.rentplace.dtos;
+package kattsyn.dev.rentplace.dtos.categories;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
